broke scrutiny_repository.go into multiple files for easier exploration & maintenance.
